Splitting up helpers, and laying the groundwork for expansion.
