+++ src/helpers/neighbors.rs
@@ -1,246 +1,7 @@
-use bevy::{
-    math::{UVec2, Vec2},
-    prelude::{BuildChildren, Color, Commands, Entity, Transform},
-};
-
-use crate::{
-    map::{HexCoordSystem, IsoCoordSystem, TilemapId, TilemapSize, TilemapTileSize, TilemapType},
-    tiles::{TileBundle, TileColor, TilePos, TileStorage, TileTexture},
-    TilemapGridSize,
-};
-
-/// Converts a tile position into an index in a vector.
-pub fn pos_2d_to_index(tile_pos: &TilePos, size: &TilemapSize) -> usize {
-    ((tile_pos.y * size.x as u32) + tile_pos.x) as usize
-}
-
-/// Calculates the position of the bottom-left of a chunk with the specified position.
-///
-/// This calculation is mostly used internally for rendering but it might be helpful so it's exposed here.
-pub fn get_chunk_2d_transform(
-    chunk_position: UVec2,
-    chunk_size: UVec2,
-    z_index: u32,
-    grid_size: Vec2,
-    map_type: &TilemapType,
-) -> Transform {
-    // Get the position of the bottom left tile of the chunk: the "anchor tile".
-    let anchor_tile_pos = TilePos {
-        x: chunk_position.x * chunk_size.x,
-        y: chunk_position.y * chunk_size.y,
-    };
-    let grid_size: TilemapGridSize = grid_size.into();
-    // Now get the position of the anchor tile.
-    let r = get_tile_pos_in_world_space(&anchor_tile_pos, &grid_size, map_type);
-    Transform::from_xyz(r.x, r.y, z_index as f32)
-}
-
-/// Returns the bottom-left coordinate of the tile associated with the specified `tile_pos`.
-pub fn get_tile_pos_in_world_space(
-    tile_pos: &TilePos,
-    grid_size: &TilemapGridSize,
-    tilemap_type: &TilemapType,
-) -> Vec2 {
-    let tile_pos_f32: Vec2 = tile_pos.into();
-    let grid_size: Vec2 = grid_size.into();
-    let mut pos = Vec2::new(grid_size.x * tile_pos_f32.x, grid_size.y * tile_pos_f32.y);
-
-    match tilemap_type {
-        TilemapType::Hexagon(HexCoordSystem::Row) => {
-            let x_offset = tile_pos_f32.y * (0.5 * grid_size.x).floor();
-            let y_offset = -1.0 * tile_pos_f32.y * (0.25 * grid_size.y).ceil();
-            pos.x += x_offset;
-            pos.y += y_offset;
-        }
-        TilemapType::Hexagon(HexCoordSystem::RowEven) => {
-            let offset = (0.25 * grid_size.x).floor();
-            if tile_pos.y % 2 == 0 {
-                pos.x -= offset;
-            } else {
-                pos.x += offset;
-            }
-            pos.y -= tile_pos_f32.y * (0.25 * grid_size.y as f32).ceil();
-        }
-        TilemapType::Hexagon(HexCoordSystem::RowOdd) => {
-            let offset = (0.25 * grid_size.x).floor();
-            if tile_pos.y % 2 == 0 {
-                pos.x += offset;
-            } else {
-                pos.x -= offset;
-            }
-            pos.y -= tile_pos_f32.y * (0.25 * grid_size.y).ceil();
-        }
-        TilemapType::Hexagon(HexCoordSystem::Column) => {
-            let x_offset = -1.0 * tile_pos_f32.x * (0.25 * grid_size.x).floor();
-            let y_offset = tile_pos_f32.x * (0.5 * grid_size.y).ceil();
-            pos.x += x_offset;
-            pos.y += y_offset;
-        }
-        TilemapType::Hexagon(HexCoordSystem::ColumnEven) => {
-            let offset = (0.25 * grid_size.y).floor();
-            if tile_pos.x % 2 == 0 {
-                pos.y -= offset;
-            } else {
-                pos.y += offset;
-            }
-            pos.x -= tile_pos_f32.x * (0.25 * grid_size.x as f32).ceil();
-        }
-        TilemapType::Hexagon(HexCoordSystem::ColumnOdd) => {
-            let offset = (0.25 * grid_size.y).floor();
-            if tile_pos.x % 2 == 0 {
-                pos.y += offset;
-            } else {
-                pos.y -= offset;
-            }
-            pos.x -= tile_pos_f32.x * (0.25 * grid_size.x).ceil();
-        }
-        TilemapType::Isometric {
-            coord_system: IsoCoordSystem::Diamond,
-            ..
-        } => {
-            pos = project_iso_diamond(tile_pos_f32.x, tile_pos_f32.y, grid_size.x, grid_size.y);
-        }
-        TilemapType::Isometric {
-            coord_system: IsoCoordSystem::Staggered,
-            ..
-        } => {
-            pos = project_iso_staggered(tile_pos_f32.x, tile_pos_f32.y, grid_size.x, grid_size.y);
-        }
-        TilemapType::Square { .. } => {}
-    };
-    pos
-}
-
-/// Fills an entire tile storage with the given tile.
-pub fn fill_tilemap(
-    tile_texture: TileTexture,
-    size: TilemapSize,
-    tilemap_id: TilemapId,
-    commands: &mut Commands,
-    tile_storage: &mut TileStorage,
-) {
-    for x in 0..size.x {
-        for y in 0..size.y {
-            let tile_pos = TilePos { x, y };
-            let tile_entity = commands
-                .spawn()
-                .insert_bundle(TileBundle {
-                    position: tile_pos,
-                    tilemap_id,
-                    texture: tile_texture,
-                    ..Default::default()
-                })
-                .id();
-            commands.entity(tilemap_id.0).add_child(tile_entity);
-            tile_storage.set(&tile_pos, Some(tile_entity));
-        }
-    }
-}
-
-/// Fills a rectangular region with the given tile.
-///
-/// The rectangular region is defined by an `origin` in `TilePos`, and a size
-/// in tiles (`TilemapSize`).  
-pub fn fill_tilemap_rect(
-    tile_texture: TileTexture,
-    origin: TilePos,
-    size: TilemapSize,
-    tilemap_id: TilemapId,
-    commands: &mut Commands,
-    tile_storage: &mut TileStorage,
-) {
-    for x in 0..size.x {
-        for y in 0..size.y {
-            let tile_pos = TilePos {
-                x: origin.x + x,
-                y: origin.y + y,
-            };
-
-            let tile_entity = commands
-                .spawn()
-                .insert_bundle(TileBundle {
-                    position: tile_pos,
-                    tilemap_id,
-                    texture: tile_texture,
-                    ..Default::default()
-                })
-                .id();
-            tile_storage.set(&tile_pos, Some(tile_entity));
-        }
-    }
-}
-
-/// Fills a rectangular region with colored versions of the given tile.
-///
-/// The rectangular region is defined by an `origin` in `TilePos`, and a size
-/// in tiles (`TilemapSize`).  
-pub fn fill_tilemap_rect_color(
-    tile_texture: TileTexture,
-    origin: TilePos,
-    size: TilemapSize,
-    color: Color,
-    tilemap_id: TilemapId,
-    commands: &mut Commands,
-    tile_storage: &mut TileStorage,
-) {
-    for x in 0..size.x {
-        for y in 0..size.y {
-            let tile_pos = TilePos {
-                x: origin.x + x,
-                y: origin.y + y,
-            };
-
-            let tile_entity = commands
-                .spawn()
-                .insert_bundle(TileBundle {
-                    position: tile_pos,
-                    tilemap_id,
-                    texture: tile_texture,
-                    color: TileColor(color),
-                    ..Default::default()
-                })
-                .id();
-            tile_storage.set(&tile_pos, Some(tile_entity));
-        }
-    }
-}
-
-/// Calculates a tilemap's centered position.
-pub fn get_centered_transform_2d(
-    size: &TilemapSize,
-    tile_size: &TilemapTileSize,
-    z_index: f32,
-) -> Transform {
-    Transform::from_xyz(
-        -(size.x as f32 * tile_size.x as f32) / 2.0,
-        -(size.y as f32 * tile_size.y as f32) / 2.0,
-        z_index,
-    )
-}
-
-/// Projects a 2D screen space point into isometric diamond space.
-///
-/// `grid_width` and `grid_height` are the dimensions of the grid in pixels.
-pub fn project_iso_diamond(x: f32, y: f32, grid_width: f32, grid_height: f32) -> Vec2 {
-    let dx = grid_width / 2.0;
-    let dy = grid_height / 2.0;
-
-    let new_x = (x + y) * dx;
-    let new_y = (-x + y) * dy;
-    Vec2::new(new_x, new_y)
-}
-
-/// Projects a 2D screen space point into isometric staggered space.
-///
-/// `grid_width` and `grid_height` are the dimensions of the grid in pixels.
-pub fn project_iso_staggered(x: f32, y: f32, grid_width: f32, grid_height: f32) -> Vec2 {
-    let dx = grid_width / 2.0;
-    let dy = grid_height / 2.0;
-
-    let new_x = x * grid_width + y * dx;
-    let new_y = y * dy;
-    Vec2::new(new_x, new_y)
-}
+use crate::map::{HexCoordSystem, IsoCoordSystem};
+use crate::tiles::TilePos;
+use crate::{TileStorage, TilemapSize, TilemapType};
+use bevy::prelude::Entity;
 
 #[derive(Clone, Copy, Debug)]
 pub enum NeighborDirection {
